% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate one stepped wedge dataset}
\usage{
generate_dataset(
  alpha,
  tau,
  theta,
  d,
  n_clusters,
  n_time_points,
  n_ind_per_cluster,
  data_type,
  sigma = NA
)
}
\arguments{
\item{alpha}{Log baseline prevalence}

\item{tau}{Cluster random effect SD}

\item{theta}{Treatment effect}

\item{d}{Treatment lag rate parameter}

\item{n_clusters}{Total number of clusters}

\item{n_time_points}{Total number of time points ("J")}

\item{n_ind_per_cluster}{Total number of individuals per cluster (assumes}

\item{data_type}{Type of outcome; options include "binomial" or "normal"}

\item{sigma}{Standard deviation of the outcome (if data_type="normal"; omit
otherwise)
equal cluster sizes)}
}
\value{
A list containing the following: \cr
    * `params`: a list of the parameters supplied in the function call \cr
    * `data`: the resulting data frame
}
\description{
Generate one stepped wedge dataset
}
