% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_spline.R
\name{log_lik_spline}
\alias{log_lik_spline}
\title{Return the log likelihood for the spline model}
\usage{
log_lik_spline(
  sigma_v,
  sigma_e,
  alpha,
  beta_1,
  beta_2,
  beta_3,
  beta_4,
  beta_5,
  theta,
  p_x,
  p_y,
  g_x,
  data
)
}
\arguments{
\item{sigma_v}{!!!!! TO DO}
}
\value{
!!!!! TO DO
}
\description{
Return the log likelihood for the spline model
}
