% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_simulation.R
\name{one_simulation}
\alias{one_simulation}
\title{Run a single simulation}
\usage{
one_simulation(L, C)
}
\arguments{
\item{L}{Passed via simba; list of simulation levels}

\item{C}{Passed via simba; list of simulation constants}
}
\value{
The list created by `run_analysis`
}
\description{
Run a single simulation
}
