% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome.R
\name{plot_outcome}
\alias{plot_outcome}
\title{Plot the outcome variable by cluster over time}
\usage{
plot_outcome(data, type)
}
\arguments{
\item{data}{A dataset returned by generate_dataset()}

\item{type}{Either "no error" or "realized"; "no error" gives the actual
value of p_ij or y_ij, whereas "realized" gives the average of the y
values}
}
\value{
A ggplot2 plot object: spaghetti plot of mean cluster outcomes over
    time
}
\description{
Plot the outcome variable by cluster over time
}
