% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sw_design.R
\name{plot_sw_design}
\alias{plot_sw_design}
\title{Plot the stepped wedge design}
\usage{
plot_sw_design(data)
}
\arguments{
\item{data}{A dataset returned by generate_dataset()}
}
\value{
A ggplot2 plot object representing the design
}
\description{
Plot the stepped wedge design
}
