% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the data analysis}
\usage{
run_analysis(data, analysis_type, data_type, L, C)
}
\arguments{
\item{data}{A dataset returned by generate_dataset()}

\item{analysis_type}{A character string corresponding to one of the
following: "2S LM" (two-stage using a simple linear model in first
stage), "2S GEE EX" (two-stage using GEE in first stage; exchangeable
correlation), "2S GEE ID" (two-stage using GEE in first stage;
independence correlation), "2S LMM ML" (two-stage using LMM in first
stage; maximum likelihood), "2S LMM REML" (two-stage using LMM in first
stage; REML), "PP SPL" (the "Purple point spline" method), "FX SPL" (the
"fixed X-coordinate" spline method), "IG LM" (simple linear model that
ignores the time-lag effect), "IG GEE" (GEE that ignores the time-lag
effect)}

\item{L}{Passed via simba; list of simulation levels}

\item{C}{Passed via simba; list of simulation constants}
}
\value{
TO DO
}
\description{
Run the data analysis
}
